[statics-refactor] refactor : 통계상세화면 Activity에서 Fragment로 변경
