--- feature/src/main/java/com/hegunhee/newsimplememoapp/feature/detailStatics/DetailStaticsActivity.kt
@@ -1,70 +1,0 @@
-package com.hegunhee.newsimplememoapp.feature.detailStatics
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.activity.viewModels
-import androidx.databinding.DataBindingUtil
-import androidx.navigation.navArgs
-import com.hegunhee.newsimplememoapp.feature.R
-import com.hegunhee.newsimplememoapp.feature.databinding.ActivityDetailStaticsBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-/**
- * Adapter의 경우 다시
- */
-@AndroidEntryPoint
-class DetailStaticsActivity : AppCompatActivity() {
-
-    private val viewModel: DetaiStaticsViewModel by viewModels()
-    private lateinit var binding: ActivityDetailStaticsBinding
-//    private val adapter = MemoAdapter() { memo ->
-//        Intent(this, DetailMemoActivity::class.java).apply {
-//            putExtra("Memo", memo)
-//            this@DetailStaticsActivity.startActivity(this)
-//        }
-//    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_detail_statics)
-        binding.run {
-            viewmodel = this@DetailStaticsActivity.viewModel
-            lifecycleOwner = this@DetailStaticsActivity
-//            recyclerview.adapter = adapter
-        }
-        viewModel.initData(navArgs<DetailStaticsActivityArgs>().value.staticsArgs)
-        initObserver()
-    }
-
-    override fun onRestart() {
-        super.onRestart()
-        viewModel.setData()
-    }
-
-    private fun initObserver() {
-        initObserverData()
-        initBackButtonObserve()
-    }
-
-    private fun initObserverData() = viewModel.detailStaticsState.observe(this) {
-//        when (it) {
-//            DetailStaticsState.Uninitialized -> {}
-//            is DetailStaticsState.Success -> {
-//                adapter.submitList(it.data)
-//            }
-//            DetailStaticsState.NullOrEmpty -> {
-//
-//            }
-//        }
-    }
-
-    private fun initBackButtonObserve() = viewModel.backButton.observe(this) {
-        when (it) {
-            BackButtonState.Back -> {
-                onBackPressed()
-            }
-            BackButtonState.Uninitialized -> {
-            }
-        }
-    }
-}
