+++ feature/src/main/res/layout/fragment_detail_statics.xml
